refactor(mobile): hoist FaviconImage/WebResultRow/NewsResultCard to primitives
