+++ apps/mobile/src/features/chat/tool-data/cards/search-results-card.tsx
@@ -1,9 +1,4 @@
-import type {
-  ImageResult,
-  NewsResult,
-  SearchResults,
-  WebResult,
-} from "@gaia/shared";
+import type { ImageResult, SearchResults, WebResult } from "@gaia/shared";
 import { BottomSheetScrollView } from "@gorhom/bottom-sheet";
 import { useEffect, useState } from "react";
 import { Image, Linking, Pressable, ScrollView, View } from "react-native";
@@ -15,69 +10,17 @@ import Animated, {
   withSequence,
   withTiming,
 } from "react-native-reanimated";
-import {
-  AppIcon,
-  Globe02Icon,
-  News01Icon,
-  Search01Icon,
-} from "@/components/icons";
+import { Search01Icon } from "@/components/icons";
 import { Text } from "@/components/ui/text";
 import {
+  FaviconImage,
+  NewsResultCard,
   ToolCardHeader,
   ToolCardInner,
   ToolCardShell,
+  WebResultRow,
 } from "@/features/chat/tool-data/primitives";
 import { BottomSheet } from "@/shared/components/ui/bottom-sheet";
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-function getHostname(url?: string): string {
-  if (!url) return "";
-  try {
-    return new URL(url).hostname;
-  } catch {
-    return url;
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Favicon image with globe fallback
-// Mirrors web's `next/image` favicon with onError → display:none. We render a
-// zinc-700 circle behind the favicon so failed loads still match the web
-// stacked-circle look.
-// ---------------------------------------------------------------------------
-
-function FaviconImage({ url, size = 14 }: { url?: string; size?: number }) {
-  const [errored, setErrored] = useState(false);
-  const hostname = getHostname(url);
-
-  if (!hostname || errored) {
-    return (
-      <View
-        className="rounded-full bg-zinc-700 items-center justify-center"
-        style={{ width: size, height: size }}
-      >
-        <AppIcon
-          icon={Globe02Icon}
-          size={Math.round(size * 0.65)}
-          color="#a1a1aa"
-        />
-      </View>
-    );
-  }
-
-  return (
-    <Image
-      source={{
-        uri: `https://www.google.com/s2/favicons?domain=${hostname}&sz=64`,
-      }}
-      style={{ width: size, height: size, borderRadius: size / 2 }}
-      onError={() => setErrored(true)}
-    />
-  );
-}
 
 // ---------------------------------------------------------------------------
 // SourcesPill — stacked favicons + "Search Results" label
@@ -146,59 +89,6 @@ function SourcesPill({
 }
 
 // ---------------------------------------------------------------------------
-// WebResultRow — used inside the bottom-sheet web list
-// Mirrors web's WebResults list item: title (sm font-medium, 1 line),
-// snippet (xs foreground-500, 2 lines), favicon + hostname row (xs primary).
-// Bottom border per row (zinc-700 / 15% white in dark).
-// ---------------------------------------------------------------------------
-
-function WebResultRow({
-  result,
-  isLast,
-}: {
-  result: WebResult;
-  isLast: boolean;
-}) {
-  const hostname = getHostname(result.url);
-  const description = result.content || result.snippet;
-
-  return (
-    <Pressable
-      onPress={() => result.url && Linking.openURL(result.url)}
-      android_ripple={{ color: "rgba(255,255,255,0.05)", borderless: false }}
-      style={{
-        paddingHorizontal: 16,
-        paddingTop: 16,
-        paddingBottom: 12,
-        borderBottomWidth: isLast ? 0 : 1,
-        borderBottomColor: "rgba(228,228,231,0.15)",
-      }}
-    >
-      <View className="gap-1">
-        <Text className="text-zinc-100 text-sm font-medium" numberOfLines={1}>
-          {result.title || hostname || "Untitled"}
-        </Text>
-
-        {!!description && (
-          <Text className="text-zinc-400 text-xs" numberOfLines={2}>
-            {description}
-          </Text>
-        )}
-
-        {!!hostname && (
-          <View className="flex-row items-center gap-2 mt-1">
-            <FaviconImage url={result.url} size={14} />
-            <Text className="text-[#00bbff] text-xs" numberOfLines={1}>
-              {hostname}
-            </Text>
-          </View>
-        )}
-      </View>
-    </Pressable>
-  );
-}
-
-// ---------------------------------------------------------------------------
 // WebResultsSheet — bottom-sheet popover equivalent
 // Mirrors web's PopoverContent → WebResults: rounded-2xl bg-zinc-800,
 // scrollable list of WebResultRow. Bottom-sheet handles the "popover" UX.
@@ -247,49 +137,6 @@ function WebResultsSheet({
         </BottomSheet.Content>
       </BottomSheet.Portal>
     </BottomSheet>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// NewsResultCard — full-bleed bg-zinc-800 cards stacked vertically
-// Mirrors web's NewsResults: rounded-lg bg-zinc-800 p-4, news icon + title
-// (text-lg font-medium, truncated 1 line, primary color), 2-line content
-// snippet (sm foreground-700), score line.
-// ---------------------------------------------------------------------------
-
-function NewsResultCard({ article }: { article: NewsResult }) {
-  return (
-    <Pressable
-      onPress={() => article.url && Linking.openURL(article.url)}
-      android_ripple={{ color: "rgba(255,255,255,0.05)", borderless: false }}
-      style={{
-        backgroundColor: "#27272a", // zinc-800
-        borderRadius: 8,
-        padding: 16,
-      }}
-    >
-      <View className="flex-row items-center gap-2 mb-1">
-        <AppIcon icon={News01Icon} size={20} color="#00bbff" />
-        <Text
-          className="text-[#00bbff] text-lg font-medium flex-1"
-          numberOfLines={1}
-        >
-          {article.title || "Untitled"}
-        </Text>
-      </View>
-
-      {!!article.content && (
-        <Text className="text-zinc-300 text-sm mb-1" numberOfLines={2}>
-          {article.content}
-        </Text>
-      )}
-
-      {typeof article.score === "number" && (
-        <Text className="text-zinc-500 text-xs">
-          Score: {article.score.toFixed(2)}
-        </Text>
-      )}
-    </Pressable>
   );
 }
