+++ apps/mobile/src/features/chat/tool-data/primitives/index.ts
@@ -3,3 +3,9 @@ export { SectionLabel } from "./section-label";
 export { ToolCardHeader } from "./tool-card-header";
 export { ToolCardInner } from "./tool-card-inner";
 export { ToolCardShell } from "./tool-card-shell";
+export {
+  FaviconImage,
+  getHostname,
+  NewsResultCard,
+  WebResultRow,
+} from "./web-result-primitives";
