+++ apps/mobile/src/uniwind-types.d.ts
@@ -1,10 +1,10 @@
 // NOTE: This file is generated by uniwind and it should not be edited manually.
 /// <reference types="uniwind/types" />
 
-declare module "uniwind" {
-  export interface UniwindConfig {
-    themes: readonly ["light", "dark"];
-  }
+declare module 'uniwind' {
+    export interface UniwindConfig {
+        themes: readonly ['light', 'dark']
+    }
 }
 
-export {};
+export {}
